% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_1.R
\name{team_1}
\alias{team_1}
\title{The team_1 function allows to create a data frame from a geometry variable}
\usage{
team_1(file, tolerance)
}
\arguments{
\item{file}{is a file path to a shape file}

\item{tolerance}{is value used for thinning the polygon}
}
\value{
df.oz.purr
}
\description{
The team_1 function allows to create a data frame from a geometry variable
}
\author{
Gani Agadilov
}
