% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_10.R
\name{team_10}
\alias{team_10}
\title{team_10
Reads in a shapefile of any country, sets a tolerance,
and returns a dataframe with information on longitude, lattitude,
group, id, and other specific information.}
\usage{
team_10(file, tolerance)
}
\arguments{
\item{file}{file path to a shapefile}

\item{tolerance}{tolerance}
}
\value{
geo_info: dataframe containing geographic and additional
information on polygons
}
\description{
team_10
Reads in a shapefile of any country, sets a tolerance,
and returns a dataframe with information on longitude, lattitude,
group, id, and other specific information.
}
\author{
Gulzina Kuttubekova
}
