% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team_11.R
\name{team_11}
\alias{team_11}
\title{Team 11 Function}
\usage{
team_11(file, tolerance)
}
\arguments{
\item{file}{input shape file datsest (.shp)}

\item{tolerance}{the value for thining the shape file. The tolerance is a single numeric value (e.g., 0.1)}
}
\description{
Function based on team_11 in lab2
Converts a shape file into a dataframe.
}
\examples{
# A shape file for Austrilia is already stored in the package called (Aus)

# Use the team_11 function to convert the shape file to a dataframe
Aus_datfra <- team_11(Aus, 0.01)
head(Aus_datfra)

}
