% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_3.R
\name{team_3}
\alias{team_3}
\title{Team 3 converting shape data file function
Function based on team 3 - lab 2 - STAT 585 2019
converts map data into a dataframe
function put together to the team package by Joshua Budi}
\usage{
team_3(file, tolerance)
}
\arguments{
\item{file}{file path of the map data}

\item{tolerance}{the unit value for thinning spatial map data (.shp)}
}
\description{
Team 3 converting shape data file function
Function based on team 3 - lab 2 - STAT 585 2019
converts map data into a dataframe
function put together to the team package by Joshua Budi
}
